requirements.txt
pytest==9.0.2
pytest-cov==7.0.0
python-dotenv==1.2.1
streamlit==1.54.0
PyMuPDF==1.27.1
Pillow==12.1.1
numpy==2.4.2
pandas==2.3.3
openpyxl==3.1.5
pdfplumber==0.11.9
fastapi==0.128.8
python-multipart==0.0.22
uvicorn==0.40.0
gunicorn==25.0.3
requests==2.32.5
pydantic==2.12.5
slowapi==0.1.9

# Gemini API (required for GEMINI_ENABLED=1)
google-genai==1.62.0

# Optional: Google Cloud Discovery Engine (for legal citations)
# Install only if VERTEX_SEARCH_ENABLED=1 is set:
# google-cloud-discoveryengine>=0.11.0

# Optional: Google Cloud Document AI (for PDF extraction when USE_DOCAI=1)
google-cloud-documentai==3.9.0
